# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IrForge - Telegram Bot Builder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans/subscribe:
    post:
      operationId: subscribeToPlan
      tags: [plans]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPlan"

  # Themes
  /themes:
    get:
      operationId: listThemes
      tags: [themes]
      summary: List available themes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"
    post:
      operationId: createTheme
      tags: [themes]
      summary: Create a custom theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
  /themes/{themeId}:
    patch:
      operationId: updateTheme
      tags: [themes]
      summary: Update a theme
      parameters:
        - name: themeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
    delete:
      operationId: deleteTheme
      tags: [themes]
      summary: Delete a theme
      parameters:
        - name: themeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /themes/apply:
    post:
      operationId: applyTheme
      tags: [themes]
      summary: Apply a theme as the active theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyThemeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"

  # Admin
...
components:
  schemas:
...
    UserPlan:
      type: object
      required: [planId, planName, status, expiresAt]
      properties:
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
          enum: [active, expired, trial, cancelled]
        expiresAt:
          type: ["string", "null"]
        renewsAt:
          type: ["string", "null"]
    SubscribeInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string

    # Themes
    Theme:
      type: object
      required: [id, name, mode, primaryColor, backgroundColor, userId, isDefault, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
        userId:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
    ThemeInput:
      type: object
      required: [name, mode, primaryColor]
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
    ThemeUpdate:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
    ApplyThemeInput:
      type: object
      required: [themeId]
      properties:
        themeId:
          type: string

    # Admin
...